requirements.txt
jupyter
shiny
altair
matplotlib
vl-convert-python
plotnine
pandas
plotly
polars
narwhals
palmerpenguins
shinylive
shinywidgets
keplergl
setuptools<81
querychat
ibis-framework
duckdb
pyarrow
pytest
pytest-playwright